test: improve coverage to ~80% with meaningful tests
- Add comprehensive governance rule evaluation tests (30 tests)
- Add scheduler unit tests for batch jobs and weekly reports (11 tests)
- Add postgres integration tests for roles, policies, events (12 tests)
- Add sync pointer and error tests (7 tests)
- Add GovernanceEvent::tenant_id() test for all variants
- Fix Role enum case-insensitive parsing (strum attribute)
- Fix postgres add_unit_policy i64/i32 type mismatch
- Fix redis subscriber to read from stream beginning

Coverage: 72.70% -> ~80% (+7.3%)

+++ storage/src/postgres.rs
@@ -465,7 +465,7 @@ impl PostgresBackend {
         unit_id: &str,
         policy: &mk_core::types::Policy,
     ) -> Result<(), PostgresError> {
-        let exists: Option<(i64,)> =
+        let exists: Option<(i32,)> =
             sqlx::query_as("SELECT 1 FROM organizational_units WHERE id = $1 AND tenant_id = $2")
                 .bind(unit_id)
                 .bind(ctx.tenant_id.as_str())
